[FLINK-23084] Move SecurityManager to flink-core
